backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
paho-mqtt==2.1.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
python-dotenv==1.0.1
